修改：图表绑定模型
