+++ mock/index.js
@@ -15,6 +15,7 @@ import search from "./search";
 import detail from "./detailMock";
 import personal from "./personal";
 import form from "./form";
+import dashboard from "./dashboard";
 
 export default {
   ...random,
@@ -22,5 +23,5 @@ export default {
   ...detail,
   ...personal,
   ...form,
-
+  ...dashboard
 };
