+++ portal/pages/dashboard/charts/sales.js
@@ -7,21 +7,13 @@
 import React, { Component } from "react";
 import G2 from "@antv/g2";
 
-const data = [
-  { type: "家用电器", value: 4544 },
-  { type: "食用酒水", value: 3321 },
-  { type: "个护健康", value: 3113 },
-  { type: "服饰箱包", value: 2341 },
-  { type: "母婴产品", value: 1231 },
-  { type: "其他", value: 1231 }
-];
-
 export default class Demo extends Component<any> {
   render() {
     return <div id="salesbox" />;
   }
 
   componentDidMount() {
+    const { data = [] } = this.props;
     const chart = new G2.Chart({
       id: "salesbox",
       forceFit: true,
