+++ portal/pages/dashboard/charts/searchusers.js
@@ -7,31 +7,13 @@
 import React, { Component } from "react";
 import G2 from "@antv/g2";
 
-const searchusersData = [
-  { time: "2019-10-15", count: 2 },
-  { time: "2019-10-16", count: 4 },
-  { time: "2019-10-17", count: 6 },
-  { time: "2019-10-18", count: 4 },
-  { time: "2019-10-19", count: 5 },
-  { time: "2019-10-20", count: 4 },
-  { time: "2019-10-21", count: 6 },
-  { time: "2019-10-22", count: 8 },
-  { time: "2019-10-23", count: 4 },
-  { time: "2019-10-24", count: 7 },
-  { time: "2019-10-25", count: 8 },
-  { time: "2019-10-26", count: 4 },
-  { time: "2019-10-27", count: 6 },
-  { time: "2019-10-28", count: 4 },
-  { time: "2019-10-29", count: 3 },
-  { time: "2019-10-30", count: 2 }
-];
-
 export default class Demo extends Component<any> {
   render() {
     return <div id="searchusers"></div>;
   }
 
   componentDidMount() {
+    const { data = [] } = this.props;
     const searchusersChart = new G2.Chart({
       id: "searchusers",
       forceFit: true,
@@ -39,7 +21,7 @@ export default class Demo extends Component<any> {
       padding: 0
     });
 
-    searchusersChart.source(searchusersData, {});
+    searchusersChart.source(data, {});
     searchusersChart.scale("time", {
       range: [0, 1]
     });
