+++ portal/pages/dashboard/charts/visits.js
@@ -7,31 +7,14 @@
 import React, { Component } from "react";
 import G2 from "@antv/g2";
 
-const visitsData = [
-  { time: "2019-10-15", count: 5 },
-  { time: "2019-10-16", count: 2 },
-  { time: "2019-10-17", count: 6 },
-  { time: "2019-10-18", count: 8 },
-  { time: "2019-10-19", count: 5 },
-  { time: "2019-10-20", count: 1 },
-  { time: "2019-10-21", count: 3 },
-  { time: "2019-10-22", count: 6 },
-  { time: "2019-10-23", count: 2 },
-  { time: "2019-10-24", count: 8 },
-  { time: "2019-10-25", count: 7 },
-  { time: "2019-10-26", count: 6 },
-  { time: "2019-10-27", count: 3 },
-  { time: "2019-10-28", count: 4 },
-  { time: "2019-10-29", count: 6 },
-  { time: "2019-10-30", count: 8 }
-];
-
 export default class Demo extends Component<any> {
   render() {
     return <div id="visits" />;
   }
 
   componentDidMount() {
+    const { data = [] } = this.props;
+
     const visitsChart = new G2.Chart({
       id: "visits",
       forceFit: true,
@@ -39,7 +22,7 @@ export default class Demo extends Component<any> {
       padding: 0
     });
 
-    visitsChart.source(visitsData, {});
+    visitsChart.source(data, {});
     visitsChart.scale("time", {
       range: [0, 1]
     });
